+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added the new Danish linguistic acceptability dataset DaLA. It's marked as 
   unofficial for now. This was added by @N-essuno ✨
+- Added the English knowledge dataset MMLU-Pro, marked as unofficial. This is a more
+  robust and challenging version of MMLU with 10 answer options per question.
 
 ## [v16.16.1] - 2026-02-25